% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_zip.R
\name{scale_fill_zip}
\alias{scale_fill_zip}
\title{Fill scale constructor for zip colors}
\usage{
scale_fill_zip(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in zip_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for zip colors
}
\examples{
ggplot(mtcars, aes(x = mpg, fill=as.factor(cyl))) +
geom_histogram(bins=8) +
theme_minimal() +
labs(title = "MPG Distribution", fill="cyl") +
scale_fill_zip()
}
