% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_zip.R
\name{zip_cols}
\alias{zip_cols}
\title{Function to extract Zip colors as hex codes}
\usage{
zip_cols(...)
}
\arguments{
\item{...}{Character names of zip_colors}
}
\description{
Function to extract Zip colors as hex codes
}
\examples{
zip_cols()
zip_cols("dark")
}
