% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_zip.R
\name{zip_pal}
\alias{zip_pal}
\title{Return function to interpolate a zip color palette}
\usage{
zip_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in zip_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a zip color palette
}
