% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_zip.R
\docType{data}
\name{zip_palettes}
\alias{zip_palettes}
\title{Zip Palettes list}
\format{
An object of class \code{list} of length 5.
}
\usage{
zip_palettes
}
\description{
Zip Palettes list
}
\keyword{datasets}
