% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_zip.R
\docType{data}
\name{zip_primary_colors}
\alias{zip_primary_colors}
\title{Zip Primary Color Palette list}
\format{
An object of class \code{list} of length 9.
}
\usage{
zip_primary_colors
}
\description{
Zip Primary Color Palette list
}
\keyword{datasets}
