% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_zip.R
\docType{data}
\name{zip_secondary_colors}
\alias{zip_secondary_colors}
\title{Zip Secondary Color Palette list}
\format{
An object of class \code{list} of length 6.
}
\usage{
zip_secondary_colors
}
\description{
Zip Secondary Color Palette list
}
\keyword{datasets}
