% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggzip.R
\name{theme_zip_classic}
\alias{theme_zip_classic}
\title{Add zip classic theme to ggplot chart}
\usage{
theme_zip_classic(
  base_size = 11,
  base_family = "Titillium Web",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\description{
This function allows you to add the zip classic theme to your ggplotgraphics.
}
\examples{
require(ggplot2)
require(showtext)
font_add_google(name = "Source Sans Pro", family = "Source Sans Pro")
showtext_auto()

ggplot(mtcars, aes(x = mpg, y = disp, color=as.factor(cyl))) +
geom_point(size=3) +
  theme_zip_classic() +
  labs(title = "Displacement vs MPG")
}
\keyword{zip_style}
